requirements.txt
Django
django-filter
djangorestframework
djangorestframework-simplejwt
drf-yasg
Pillow
psycopg2-binary
python-decouple

